% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coarse_taxa.R
\name{add_coarse_taxa}
\alias{add_coarse_taxa}
\title{Add coarse taxa for complex taxon names}
\usage{
add_coarse_taxa(data, match_file = NULL, output = "Phylum")
}
\arguments{
\item{data}{A data frame with a column named \code{Taxon}}

\item{match_file}{A data.frame with at least two columns. One of the columns
must be named \code{Taxon}; the name of the other column is not
restricted and shall be passed to the \code{output} arguement.}

\item{output}{The name of the output column.}
}
\value{

}
\description{
Add coarse taxa for complex taxon names
}
\examples{
a <- data.frame(Taxon = c("Polychaeta", "Nematoda", "Hydrozoa"))
add_coarse_taxa(data = a, match_file = NULL, output = "Phylum")
}
