% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_method.R
\name{assign_method}
\alias{assign_method}
\title{Assign volume estimation method in to each individuals}
\usage{
assign_method(data, method_file = NULL)
}
\arguments{
\item{data}{The data that records observations of each individuals size measurements
with at least the following column: "Taxon"}

\item{method_file}{The file that contains the pairwise biovolume estimation method.
Note that the method file has a specific format to follow. See
\code{biovolume method} for more information.
If NULL, the method file will use the default "biovolume_method"
data.}
}
\value{

}
\description{
This function adds the following two columns to your original data frame:
\itemize{
\item \code{Type} : This column stores the method for volume calculation.
\item \code{C} : This column stores the conversion factors for individuals that uses
length-weight relationships for biomass calculation.
}
}
\examples{
# the default estimation method for each available taxon.
biovolume_method
a <- data.frame(Taxon = c("Polychaeta", "Oligochaeta", "Sipuncula", "Not in the column"))
assign_method(a)
}
