% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_type.R
\name{assign_type}
\alias{assign_type}
\title{Assign volume estimation types to each individual}
\usage{
assign_type(data, type_file = NULL)
}
\arguments{
\item{data}{The data that records observations of each individuals size measurements
with at least the following column: "Taxon"}

\item{type_file}{The file that contains the pairwise biovolume estimation type.
Note that the type file has a specific format to follow. See
\code{biovolume type} for more information.
If NULL, the type file will use the default "biovolume_type"
data.}
}
\value{

}
\description{
This function adds the following two columns to your original data frame:
\itemize{
\item \code{Type} : This column stores the type for volume calculation.
\item \code{C} : This column stores the conversion factors for individuals that uses
length-weight relationships for biomass calculation.
}
}
\details{
The original \code{Type} column will be overwritten with the \code{Type} provided in \code{type_file}.
}
\examples{
# the default estimation type for each available taxon.
biovolume_type
a <- data.frame(Taxon = c("Polychaeta", "Oligochaeta", "Sipuncula", "Not in the column"))
assign_type(a)
}
