% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biovolume_type.R
\docType{data}
\name{biovolume_type}
\alias{biovolume_type}
\title{Biovolume type}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 44 rows and 6 columns.
}
\usage{
biovolume_type
}
\description{
This data documents the biovolume estimation methods
respective to each taxon.
\itemize{
\item \code{Taxon} Taxa in the recorded form.
\item \code{Note} Parsing the taxa into different estimation methods.
\item \code{Type} The methods for biovolume methods. Currently
allowing cone, cylinder, ellipsoid, and LWR.
\item \code{C} The conversion factors for LWR.
\item \code{C_origin} The origin of conversion factors for taxa
without published factors.
\item \code{Comment} Comment about the whole row.
}
}
\keyword{datasets}
