% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_biovolume.R
\name{calculate_biovolume}
\alias{calculate_biovolume}
\title{Calculate the biovolume of each individual with their respective body shape types}
\usage{
calculate_biovolume(data)
}
\arguments{
\item{data}{A long format data containing columns of Length (L), Width (W), and body shape type (Type).}
}
\value{
The same data frame with a new column called "Size" contains the volume data.
}
\description{
Calculates biovolume of each individual with its assigned method.
}
\details{
Note that this function will first remove all the entries in the \code{Size} column.
}
\examples{
a <- data.frame(
  L = c(10, 5, 6, 78, 1, 6),
  W = c(0.1, 3, 4, 1, 2, 3),
  Method = c("LWR", "LWR", "Cylinder", "Cylinder", "Cone", "Elliposid"),
  C = c(0.53, 0.45, NA, NA, NA, NA)
)
calculate_biovolume(a)
}
