% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ophiuroid_size.R
\name{calculate_ophiuroid_size}
\alias{calculate_ophiuroid_size}
\title{Combine of ophiuroids' arm and disc volume}
\usage{
calculate_ophiuroid_size(data, ophiuroid_method, grouping_variables)
}
\arguments{
\item{data}{A long format data that had their biovolume stored in the column "Size" and
required comments stored in the column "Note".}

\item{ophiuroid_method}{Currently have two options:.
\itemize{
\item \code{all_arms}: calculate the biovolume of one ophiuroid by simply add all the arms and
and their respective disc.
\item \code{longest_arms} : calculate the biovolume of one ophiuroid by seeking the longest arm
that is still attached to the disc and assume all arms are of the
same length as the longest arms.
}}

\item{grouping_variables}{A character vector of the subset of column names. The variables
should at least include the dates, stations, and replicates to
prevent individuals across different samples from being added
together.}
}
\value{
The modified input data with the size of the ophiuroids manipulated.
Length and width will be left as NA. Conditions of all individuals are
noted as complete (C).
}
\description{
Note that this function only works under my measurement
protocol for ophiuroids. If you use different measurement protocols
while using this function, dubious results might appear. See
example for more insight.
}
\examples{
a <- data.frame(
  Taxon = rep("Ophiuroidea", 6),
  Size = runif(6, min = 1, max = 5),
  Note = c("Dics-1", rep("Arm-1", 4), "Arm")
)
calculate_ophiuroid_size(data = a, protocol = "all_arms")
calculate_ophiuroid_size(data = a, protocol = "longest_arm")
}
