% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coarse_taxa.R
\docType{data}
\name{coarse_taxa}
\alias{coarse_taxa}
\title{Coarse taxa}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 37 rows and 3 columns.
}
\usage{
coarse_taxa
}
\description{
This data documents the respective phylum and groupings for the sake of concise graphs

\itemize{
\item \code{Taxon} Taxa in the recorded form.
\item \code{Phylum} The phylum each taxon is currently in.
\item \code{Group} The taxonomic group arbitrarily defined for concise graphing.
}
}
\keyword{datasets}
