% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cone.R
\name{cone}
\alias{cone}
\title{Estimate biovolume with cone equation}
\usage{
cone(Length, Width)
}
\arguments{
\item{Length}{Length of an organism}

\item{Width}{Width of an organism}
}
\value{

}
\description{
Estimate the biovolume of individuals with cone equation.
}
\examples{
cone(30, 1)
}
