% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\docType{data}
\name{ctd}
\alias{ctd}
\title{CTD downcast profiles}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 540 rows and 14 columns.
}
\usage{
ctd
}
\description{
CTD downcast profiles
}
\details{
The condition and unit of each column were listed below.
\itemize{
\item{\code{Cruise}} The cruise during sampling.
\item{\code{Habitat}} Habitat type (shelf, slope, canyon)
\item{\code{Station}} Sampling station.
\item{\code{Date}} Date recorded in YYYY-MM-DD.
\item{\code{Latitude}} Latitude in decimals.
\item{\code{Longitude}} Longitude in decimals.
\item{\code{Pressure}} Pressure in decibar.
\item{\code{Temperature}} Temperature in Celcius.
\item{\code{Salinity}} Salinity in Practical salinity unit (PSU).
\item{\code{SigmaTheta}} Potential density anomaly (Sigma-Theta) in kg/m^3.
\item{\code{Density}} Potential density in kg/m^3.
\item{\code{Oxygen}} Dissolved oxygen in mg/L.
\item{\code{Fluorescence}} Estimated chlorophyll concentration in mg/L.
\item{\code{Transmission}} Light transmission in percent.
}
}
\author{
Yen-Ting Chen \email{r08241220@ntu.edu.tw}
}
\keyword{datasets}
