% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cylinder.R
\name{cylinder}
\alias{cylinder}
\title{Estimate biovolume with cylinder equation}
\usage{
cylinder(Length, Width)
}
\arguments{
\item{Length}{Length of an organism}

\item{Width}{Width of an organism}
}
\value{
The volume
}
\description{
Estimate the biovolume of individuals with cylinder equation.
}
\examples{
cylinder(15, 0.75)
}
