% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid.R
\name{ellipsoid}
\alias{ellipsoid}
\title{Estimate biovolume with ellipsoid equation}
\usage{
ellipsoid(Length, Width)
}
\arguments{
\item{Length}{Length of an organism}

\item{Width}{Width of an organism}
}
\value{

}
\description{
Estimate the biovolume of individuals with ellipsoid equation.
}
\examples{
ellipsoid(10, 5)
}
