% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwr.R
\name{lwr}
\alias{lwr}
\title{Estimate biovolume with Length-width Relationship}
\usage{
lwr(Length, Width, Conversion_factor)
}
\arguments{
\item{Length}{Length of an organism}

\item{Width}{Width of an organism}

\item{Conversion_factor}{Shapeless conversion factor. For available conversion factors,
please see Fellwe & Warwick (1998).}
}
\value{

}
\description{
Estimate the biovolume of individuals with a published shapeless conversion factors.
}
\examples{
lwr(10, 20, 0.5)
}
