% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{map}
\alias{map}
\title{Mapping data}
\description{
The longitude, latitude, and elevation of the study area.
The object \code{bathy_map} was extracted from GEBCO2020.
}
\details{
The condition and unit of each column were listed below:
\itemize{
\item{\code{Longitude}}
\item{\code{Latitude}}
\item{\code{Elevation}}
}
}
\references{

}
\author{
Yen-Ting Chen \email{r08241220@ntu.edu.tw}
}
\keyword{datasets}
