% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ou.R
\docType{data}
\name{ou}
\alias{ou}
\title{Oxygen utilization data}
\format{
An object of class \code{data.frame} with 33 rows and 8 columns.
}
\usage{
ou
}
\description{
Oxygen utilization data
}
\details{
The condition and unit of each column were listed below.
\itemize{
\item{\code{Cruise}} The cruise during sampling.
\item{\code{Station}} Sampling station.
\item{\code{Deployment}} Number of deployment.
\item{\code{Tube}} Order of tubes.
\item{\code{TOU}} Total oxygen utilization of the sediment (mmol/m2/d).
\item{\code{DOU}} Dissolved oxygen utilzation of the sediment (mmol/m2/d).
\item{\code{BOU}} Benthic-meidated oxygen utilization of the sediment. Derived by subtracting TOU with DOU.
\item{\code{OPD}} Oxygen penetration depth (cm).
}
}
\author{
Yen-Ting Chen \email{r08241220@ntu.edu.tw}
}
\keyword{datasets}
