% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizedata.R
\docType{data}
\name{sizedata}
\alias{sizedata}
\alias{OR1_1219_macro_size}
\alias{OR1_1242_macro_size}
\title{Individual size data of all cruises}
\format{
An object of class \code{data.frame} with 1329 rows and 14 columns.

An object of class \code{data.frame} with 1271 rows and 14 columns.
}
\usage{
OR1_1219_macro_size

OR1_1242_macro_size
}
\description{
Individual size data of all cruises
}
\details{
The condition and unit of each column were listed below.
\itemize{
\item{\code{Cruise}} The cruise during sampling.
\item{\code{Habitat}} Habitat type (Shelf, slope, canyon)
\item{\code{Station}} Sampling station.
\item{\code{Deployment}} Number of deployment.
\item{\code{Tube}} Order of tubes.
\item{\code{Section}} sampled section in cm
\item{\code{Taxon}} Recorded grouping taxa of the specimen
\item{\code{Condition}} The condition of the specimen
\item{\code{L}} maximum body length in mm
\item{\code{W}} maximum body length in mm
\item{\code{size}} the volume of the specimen in mm^3
\item{\code{Note}} Comments and further information for the specimen
\item{\code{Type}} The respective geometric shapes for biovolume estimation
\item{\code{C}} The conversion factors for biovolume estimation
\item{\code{WM}} wet mass in mg
}
}
\author{
Yen-Ting Chen \email{r08241220@ntu.edu.tw}
}
\keyword{datasets}
