% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ou.R
\docType{data}
\name{ou}
\alias{ou}
\alias{tou}
\alias{dou}
\title{Sediment oxygen community utilization data}
\format{
An object of class \code{data.frame} with 33 rows and 10 columns.

An object of class \code{data.frame} with 94 rows and 11 columns.
}
\usage{
tou

dou
}
\description{
Sediment oxygen community utilization data
}
\details{
The condition and unit of each column were listed below.
\itemize{
\item{\code{Cruise}} The cruise during sampling.
\item{\code{Habitat}} Habitat type (shelf, slope, canyon)
\item{\code{Station}} Sampling station.
\item{\code{Date}} Date recorded in YYYY-MM-DD.
\item{\code{Deployment}} Number of deployment.
\item{\code{Tube}} Order of tubes.
\item{\code{Channel}} Microelectrode channels; only in \code{dou} data set.
\item{\code{Incub_temperature}} Incubated water temperature in Celcius.
\item{\code{Incub_TOU}} Incubated total oxygen utilization in mmol/m^2/d^1.
\item{\code{Incub_DOU}} Incubated dissolved oxygen utilization in mmol/m^2/d^1.
\item{\code{In_situ_temperature}} Bottom water temperature in Celcius.
\item{\code{In_situ_TOU}} In situ total oxygen utilization in mmol/m^2/d^1.
\item{\code{In_situ_DOU}} In situ dissolved oxygen utilization in mmol/m^2/d^1.
\item{\code{OPD}} Oxygen penetration depth in cm.
}
}
\author{
Yen-Ting Chen \email{r08241220@ntu.edu.tw}
}
\keyword{datasets}
